+++ src/entrypoint/handler.h
@@ -16,7 +16,8 @@ public:
     coro<void> handle(boost::asio::ip::tcp::socket s) override;
 
     coro<http::response> handle_request(boost::asio::ip::tcp::socket& s,
-                                        http::request& req);
+                                        http::request& req,
+                                        const std::string& id);
 
 private:
     command_factory m_command_factory;